+++ src/AttributeResolverFactoryInterface.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Hydrator;
 
-use Yiisoft\Hydrator\DataAttributeInterface;
 use Yiisoft\Hydrator\Exception\NonInstantiableException;
-use Yiisoft\Hydrator\ParameterAttributeInterface;
 
 interface AttributeResolverFactoryInterface
 {
