+++ src/ParameterAttributeResolverInterface.php
@@ -13,9 +13,12 @@ interface ParameterAttributeResolverInterface
      * Returns the resolved from specified attribute value object.
      *
      * @param ParameterAttributeInterface $attribute The attribute to be resolved.
-     * @param Context $context The context of value resolving from attribute.
+     * @param ParameterAttributeResolveContext $context The context of value resolving from attribute.
      *
      * @return Result The resolved from specified attribute value object.
      */
-    public function getParameterValue(ParameterAttributeInterface $attribute, Context $context): Result;
+    public function getParameterValue(
+        ParameterAttributeInterface $attribute,
+        ParameterAttributeResolveContext $context
+    ): Result;
 }
