+++ tests/Support/Attribute/ContextViewerResolver.php
@@ -4,21 +4,21 @@ declare(strict_types=1);
 
 namespace Yiisoft\Hydrator\Tests\Support\Attribute;
 
-use Yiisoft\Hydrator\Context;
+use Yiisoft\Hydrator\ParameterAttributeResolveContext;
 use Yiisoft\Hydrator\ParameterAttributeInterface;
 use Yiisoft\Hydrator\ParameterAttributeResolverInterface;
 use Yiisoft\Hydrator\Result;
 
 final class ContextViewerResolver implements ParameterAttributeResolverInterface
 {
-    private ?Context $context = null;
+    private ?ParameterAttributeResolveContext $context = null;
 
-    public function getContext(): ?Context
+    public function getContext(): ?ParameterAttributeResolveContext
     {
         return $this->context;
     }
 
-    public function getParameterValue(ParameterAttributeInterface $attribute, Context $context): Result
+    public function getParameterValue(ParameterAttributeInterface $attribute, ParameterAttributeResolveContext $context): Result
     {
         $this->context = $context;
         return Result::fail();
